fix: resolve all clippy warnings
- pcap_file.rs: remove unnecessary i64 cast on tv_sec
- geo/mod.rs: replace redundant |r| Arc::new(r) closure with Arc::new
- rotating_pcap.rs: map_or(false, ..) -> is_some_and(..)
- rules/mod.rs: map_or(false, ..) -> is_some_and(..); collapse nested if let
- stats/mod.rs: sort_unstable_by -> sort_unstable_by_key(Reverse)
- stream/mod.rs: loop+let-else -> while let; add is_empty() for len_without_is_empty

+++ arwah-engine/src/capture/pcap_file.rs
@@ -38,7 +38,7 @@ impl CaptureSource for PcapFileCapture {
             Ok(pkt) => {
                 let ts = pkt.header.ts;
                 let timestamp = Utc
-                    .timestamp_opt(ts.tv_sec as i64, ts.tv_usec as u32 * 1000)
+                    .timestamp_opt(ts.tv_sec, ts.tv_usec as u32 * 1000)
                     .single()
                     .unwrap_or_else(Utc::now);
                 Ok(Some(RawPacket {
